move inf and nan operations to high-level float api
