+++ crates/burn-tensor/src/tensor/api/float.rs
@@ -1,10 +1,19 @@
 use crate::FloatDType;
 use crate::Tensor;
+use crate::cast::ToElement;
 use crate::quantization::{QuantScheme, QuantizationParameters};
 use crate::tensor::backend::Backend;
 use crate::tensor::stats;
 use crate::tensor::{Distribution, TensorData};
 use crate::{Int, TensorPrimitive};
+
+use super::Bool;
+
+/// Default RTOL value for `is_close` and `all_close`.
+pub const DEFAULT_RTOL: f64 = 1e-5;
+
+/// Default ATOL value for `is_close` and `all_close`.
+pub const DEFAULT_ATOL: f64 = 1e-8;
 
 impl<const D: usize, B> Tensor<B, D>
 where
@@ -385,5 +394,223 @@ where
     /// The dequantized tensor.
     pub fn dequantize(self) -> Tensor<B, D> {
         Tensor::new(TensorPrimitive::Float(self.primitive.tensor()))
+    }
+
+    /// Checks element wise if the tensor is close to another tensor.
+    ///
+    /// The tolerance is defined by the following equation:
+    ///
+    /// ```text
+    /// abs(a - b) <= (atol + rtol * abs(b))
+    ///
+    /// where `a` is the first tensor, `b` is the second tensor, `rtol` is the relative tolerance,
+    /// and `atol` is the absolute tolerance.
+    /// ```
+    ///
+    /// # Arguments
+    ///
+    /// * `other` - The tensor to compare with.
+    /// * `rtol` - Optional relative tolerance. Default is 1e-5; see `DEFAULT_RTOL`.
+    /// * `atol` - Optional absolute tolerance. Default is 1e-8; see `DEFAULT_ATOL`.
+    ///
+    /// # Returns
+    ///
+    /// A boolean tensor with the same shape as the input tensors.
+    ///
+    /// # Example
+    ///
+    /// ```rust
+    /// use burn_tensor::backend::Backend;
+    /// use burn_tensor::{Tensor, Shape};
+    ///
+    /// fn example<B: Backend>() {
+    ///    let device = B::Device::default();
+    ///    let tensor1 = Tensor::<B, 2>::from_data([[1.0, -2.0, 3.0], [5.0, 9.0, 6.0]], &device);
+    ///    let tensor2 = Tensor::<B, 2>::from_data([[1.0, -2.0, 3.0], [5.0, 9.0, 6.0]], &device);
+    ///    let tensor = tensor1.is_close(tensor2, None, None);
+    ///    println!("{tensor}");
+    ///    // [[true, true, true], [true, true, true]]
+    /// }
+    /// ```
+    pub fn is_close(self, other: Self, rtol: Option<f64>, atol: Option<f64>) -> Tensor<B, D, Bool> {
+        let rtol = rtol.unwrap_or(DEFAULT_RTOL);
+        let atol = atol.unwrap_or(DEFAULT_ATOL);
+
+        // check finite difference is close
+        let is_close_finite_val = self
+            .clone()
+            .sub(other.clone())
+            .abs()
+            .lower_equal(other.clone().abs().mul_scalar(rtol).add_scalar(atol))
+            .bool_and(self.clone().is_finite())
+            .bool_and(other.clone().is_finite());
+
+        // check if both are infinite and have same sign
+        let inf_same_sign = self
+            .clone()
+            .is_finite()
+            .bool_not()
+            .bool_and(other.clone().is_finite().bool_not())
+            .bool_and(self.equal(other));
+
+        is_close_finite_val.bool_or(inf_same_sign)
+    }
+
+    /// Checks if all elements are close to another tensor.
+    ///
+    /// The tolerance is defined by the following equation:
+    ///
+    /// ```text
+    ///
+    /// abs(a - b) <= (atol + rtol * abs(b))
+    ///
+    /// where `a` is the first tensor, `b` is the second tensor, `rtol` is the relative tolerance,
+    /// and `atol` is the absolute tolerance.
+    ///
+    /// ```
+    ///
+    /// # Arguments
+    ///
+    /// * `other` - The tensor to compare with.
+    /// * `rtol` - Optional relative tolerance. Default is 1e-5; see `DEFAULT_RTOL`.
+    /// * `atol` - Optional absolute tolerance. Default is 1e-8; see `DEFAULT_ATOL`.
+    ///
+    /// # Returns
+    ///
+    /// A boolean scalar.
+    ///
+    /// # Remarks
+    ///
+    /// # Example
+    ///
+    /// ```rust
+    /// use burn_tensor::backend::Backend;
+    /// use burn_tensor::{Tensor, Shape};
+    ///
+    /// fn example<B: Backend>() {
+    ///    let device = B::Device::default();
+    ///    let tensor1 = Tensor::<B, 2>::from_data([[1.0, -2.0, 3.0], [5.0, 9.0, 6.0]], &device);
+    ///    let tensor2 = Tensor::<B, 2>::from_data([[1.0, -2.0, 3.0], [5.0, 9.0, 6.0]], &device);
+    ///    let result = tensor1.all_close(tensor2, None, None);
+    ///    println!("{}", result);
+    ///    // true
+    /// }
+    /// ```
+    pub fn all_close(self, other: Self, rtol: Option<f64>, atol: Option<f64>) -> bool {
+        self.is_close(other, rtol, atol)
+            .all()
+            .into_scalar()
+            .to_bool()
+    }
+
+    /// Returns a new tensor with boolean elements indicating whether each element of the input is NaN.
+    ///
+    /// # Returns
+    ///
+    /// A boolean tensor where `true` indicates NaN and `false` indicates a non-NaN value.
+    ///
+    /// # Example
+    ///
+    /// ```rust
+    /// use burn_tensor::backend::Backend;
+    /// use burn_tensor::{Tensor, Bool, Shape};
+    ///
+    /// fn example<B: Backend>() {
+    ///    let device = B::Device::default();
+    ///    let tensor = Tensor::<B, 2>::from_data([[1.0, f64::NAN, 3.0], [5.0, 9.0, 6.0]], &device);
+    ///    let tensor = tensor.is_nan();
+    ///    println!("{tensor}");
+    ///    // [[false, true, false], [false, false, false]]
+    /// }
+    /// ```
+    pub fn is_nan(self) -> Tensor<B, D, Bool> {
+        // Check if the input tensor is NaN by comparing it to itself
+        // NaN is the only value that is not equal to itself
+        self.clone().not_equal(self)
+    }
+
+    /// Checks if the tensor contains any NaN values.
+    ///
+    /// # Returns
+    ///
+    /// A boolean tensor with a single element indicating whether the tensor contains any NaN values.
+    ///
+    /// # Example
+    ///
+    /// ```rust
+    /// use burn_tensor::backend::Backend;
+    /// use burn_tensor::{Tensor, Bool, Shape};
+    ///
+    /// fn example<B: Backend>() {
+    ///   let device = B::Device::default();
+    ///   let tensor = Tensor::<B, 2>::from_data([[1.0, -2.0, 3.0], [f64::NAN, 9.0, 6.0]], &device);
+    ///   let tensor = tensor.contains_nan();
+    ///   println!("{tensor}");
+    ///   // [true]
+    ///   let tensor = Tensor::<B, 2>::from_data([[1.0, -2.0, 3.0], [5.0, 9.0, 6.0]], &device);
+    ///   let tensor = tensor.contains_nan();
+    ///   println!("{tensor}");
+    ///   // [false]
+    /// }
+    /// ```
+    pub fn contains_nan(self) -> Tensor<B, 1, Bool> {
+        // Summing the tensor will result in NaN if the tensor contains any NaN values
+        // This is faster than checking each element individually
+        // because it rolls up the NaN values into a single value
+        let sum = self.sum();
+
+        sum.is_nan()
+    }
+
+    /// Returns a new tensor with boolean elements indicating whether each element of the input is infinite (either +INF or -INF).
+    ///
+    /// # Returns
+    ///
+    /// A boolean tensor where `true` indicates that the value is infinite
+    ///
+    /// # Example
+    ///
+    /// ```rust
+    /// use burn_tensor::backend::Backend;
+    /// use burn_tensor::{Tensor, Bool, Shape};
+    ///
+    /// fn example<B: Backend>() {
+    ///    let device = B::Device::default();
+    ///    let tensor = Tensor::<B, 2>::from_data([[1.0, f64::INFINITY, 3.0], [f64::NAN, 9.0, 6.0]], &device);
+    ///    let tensor = tensor.is_finite();
+    ///    println!("{tensor}");
+    ///    // [[false, true, false], [false, false, false]]
+    /// }
+    /// ```
+    pub fn is_inf(self) -> Tensor<B, D, Bool> {
+        self.abs().equal_elem(f64::INFINITY)
+    }
+
+    /// Returns a new tensor with boolean elements indicating whether each element of the input is finite
+    ///
+    /// # Returns
+    ///
+    /// A boolean tensor where `true` indicates that the value is finite and `false` indicates
+    /// either INF, -INF or NAN
+    ///
+    /// # Example
+    ///
+    /// ```rust
+    /// use burn_tensor::backend::Backend;
+    /// use burn_tensor::{Tensor, Bool, Shape};
+    ///
+    /// fn example<B: Backend>() {
+    ///    let device = B::Device::default();
+    ///    let tensor = Tensor::<B, 2>::from_data([[1.0, f64::INFINITY, 3.0], [f64::NAN, 9.0, 6.0]], &device);
+    ///    let tensor = tensor.is_finite();
+    ///    println!("{tensor}");
+    ///    // [[true, false, true], [false, true, true]]
+    /// }
+    /// ```
+    pub fn is_finite(self) -> Tensor<B, D, Bool> {
+        self.clone()
+            .is_nan()
+            .bool_not()
+            .bool_and(self.is_inf().bool_not())
     }
 }
