+++ crates/burn-tensor/src/tensor/api/mod.rs
@@ -17,6 +17,7 @@ pub use argwhere::argwhere_data;
 pub use autodiff::*;
 pub use base::*;
 pub use cartesian_grid::cartesian_grid;
+pub use float::{DEFAULT_ATOL, DEFAULT_RTOL};
 pub use kind::*;
 pub use numeric::*;
 pub use slice::*;
