+++ crates/burn-tensor/src/tensor/api/numeric.rs
@@ -1,6 +1,6 @@
 use alloc::vec::Vec;
 
-use crate::{TensorMetadata, alloc::borrow::ToOwned, cast::ToElement};
+use crate::{TensorMetadata, alloc::borrow::ToOwned};
 
 use crate::TensorPrimitive;
 use crate::quantization::QTensorPrimitive;
@@ -12,12 +12,6 @@ use crate::{
     check::TensorCheck,
     ops::{Device, IntTensor},
 };
-
-/// Default RTOL value for `is_close` and `all_close`.
-pub const DEFAULT_RTOL: f64 = 1e-5;
-
-/// Default ATOL value for `is_close` and `all_close`.
-pub const DEFAULT_ATOL: f64 = 1e-8;
 
 impl<B, const D: usize, K> Tensor<B, D, K>
 where
@@ -2119,233 +2113,6 @@ where
     pub fn matmul(self, other: Tensor<B, D, K>) -> Tensor<B, D, K> {
         check!(TensorCheck::matmul(&self, &other));
         Tensor::new(K::matmul(self.primitive, other.primitive))
-    }
-}
-
-impl<B, const D: usize> Tensor<B, D, Float>
-where
-    B: Backend,
-{
-    /// Checks element wise if the tensor is close to another tensor.
-    ///
-    /// The tolerance is defined by the following equation:
-    ///
-    /// ```text
-    /// abs(a - b) <= (atol + rtol * abs(b))
-    ///
-    /// where `a` is the first tensor, `b` is the second tensor, `rtol` is the relative tolerance,
-    /// and `atol` is the absolute tolerance.
-    /// ```
-    ///
-    /// # Arguments
-    ///
-    /// * `other` - The tensor to compare with.
-    /// * `rtol` - Optional relative tolerance. Default is 1e-5; see `DEFAULT_RTOL`.
-    /// * `atol` - Optional absolute tolerance. Default is 1e-8; see `DEFAULT_ATOL`.
-    ///
-    /// # Returns
-    ///
-    /// A boolean tensor with the same shape as the input tensors.
-    ///
-    /// # Example
-    ///
-    /// ```rust
-    /// use burn_tensor::backend::Backend;
-    /// use burn_tensor::{Tensor, Shape};
-    ///
-    /// fn example<B: Backend>() {
-    ///    let device = B::Device::default();
-    ///    let tensor1 = Tensor::<B, 2>::from_data([[1.0, -2.0, 3.0], [5.0, 9.0, 6.0]], &device);
-    ///    let tensor2 = Tensor::<B, 2>::from_data([[1.0, -2.0, 3.0], [5.0, 9.0, 6.0]], &device);
-    ///    let tensor = tensor1.is_close(tensor2, None, None);
-    ///    println!("{tensor}");
-    ///    // [[true, true, true], [true, true, true]]
-    /// }
-    /// ```
-    pub fn is_close(self, other: Self, rtol: Option<f64>, atol: Option<f64>) -> Tensor<B, D, Bool> {
-        let rtol = rtol.unwrap_or(DEFAULT_RTOL);
-        let atol = atol.unwrap_or(DEFAULT_ATOL);
-
-        // check finite difference is close
-        let is_close_finite_val = Tensor::new(Float::lower_equal(
-            Float::abs(Float::sub(self.primitive.clone(), other.primitive.clone())),
-            Float::add_scalar(
-                Float::mul_scalar(Float::abs(other.primitive.clone()), rtol),
-                atol,
-            ),
-        ))
-        .bool_and(self.is_finite())
-        .bool_and(other.is_finite());
-
-        // check if both are infinite and have same sign
-        let inf_same_sign = self
-            .is_finite()
-            .bool_not()
-            .bool_and(other.is_finite().bool_not())
-            .bool_and(self.equal(other));
-
-        is_close_finite_val.bool_or(inf_same_sign)
-    }
-
-    /// Checks if all elements are close to another tensor.
-    ///
-    /// The tolerance is defined by the following equation:
-    ///
-    /// ```text
-    ///
-    /// abs(a - b) <= (atol + rtol * abs(b))
-    ///
-    /// where `a` is the first tensor, `b` is the second tensor, `rtol` is the relative tolerance,
-    /// and `atol` is the absolute tolerance.
-    ///
-    /// ```
-    ///
-    /// # Arguments
-    ///
-    /// * `other` - The tensor to compare with.
-    /// * `rtol` - Optional relative tolerance. Default is 1e-5; see `DEFAULT_RTOL`.
-    /// * `atol` - Optional absolute tolerance. Default is 1e-8; see `DEFAULT_ATOL`.
-    ///
-    /// # Returns
-    ///
-    /// A boolean scalar.
-    ///
-    /// # Remarks
-    ///
-    /// # Example
-    ///
-    /// ```rust
-    /// use burn_tensor::backend::Backend;
-    /// use burn_tensor::{Tensor, Shape};
-    ///
-    /// fn example<B: Backend>() {
-    ///    let device = B::Device::default();
-    ///    let tensor1 = Tensor::<B, 2>::from_data([[1.0, -2.0, 3.0], [5.0, 9.0, 6.0]], &device);
-    ///    let tensor2 = Tensor::<B, 2>::from_data([[1.0, -2.0, 3.0], [5.0, 9.0, 6.0]], &device);
-    ///    let result = tensor1.all_close(tensor2, None, None);
-    ///    println!("{}", result);
-    ///    // true
-    /// }
-    /// ```
-    pub fn all_close(self, other: Self, rtol: Option<f64>, atol: Option<f64>) -> bool {
-        self.is_close(other, rtol, atol)
-            .all()
-            .into_scalar()
-            .to_bool()
-    }
-
-    /// Returns a new tensor with boolean elements indicating whether each element of the input is NaN.
-    ///
-    /// # Returns
-    ///
-    /// A boolean tensor where `true` indicates NaN and `false` indicates a non-NaN value.
-    ///
-    /// # Example
-    ///
-    /// ```rust
-    /// use burn_tensor::backend::Backend;
-    /// use burn_tensor::{Tensor, Bool, Shape};
-    ///
-    /// fn example<B: Backend>() {
-    ///    let device = B::Device::default();
-    ///    let tensor = Tensor::<B, 2>::from_data([[1.0, f64::NAN, 3.0], [5.0, 9.0, 6.0]], &device);
-    ///    let tensor = tensor.is_nan();
-    ///    println!("{tensor}");
-    ///    // [[false, true, false], [false, false, false]]
-    /// }
-    /// ```
-    pub fn is_nan(&self) -> Tensor<B, D, Bool> {
-        // Check if the input tensor is NaN by comparing it to itself
-        // NaN is the only value that is not equal to itself
-        Tensor::new(Float::not_equal(
-            self.primitive.clone(),
-            self.primitive.clone(),
-        ))
-    }
-
-    /// Checks if the tensor contains any NaN values.
-    ///
-    /// # Returns
-    ///
-    /// A boolean tensor with a single element indicating whether the tensor contains any NaN values.
-    ///
-    /// # Example
-    ///
-    /// ```rust
-    /// use burn_tensor::backend::Backend;
-    /// use burn_tensor::{Tensor, Bool, Shape};
-    ///
-    /// fn example<B: Backend>() {
-    ///   let device = B::Device::default();
-    ///   let tensor = Tensor::<B, 2>::from_data([[1.0, -2.0, 3.0], [f64::NAN, 9.0, 6.0]], &device);
-    ///   let tensor = tensor.contains_nan();
-    ///   println!("{tensor}");
-    ///   // [true]
-    ///   let tensor = Tensor::<B, 2>::from_data([[1.0, -2.0, 3.0], [5.0, 9.0, 6.0]], &device);
-    ///   let tensor = tensor.contains_nan();
-    ///   println!("{tensor}");
-    ///   // [false]
-    /// }
-    /// ```
-    pub fn contains_nan(&self) -> Tensor<B, 1, Bool> {
-        // Summing the tensor will result in NaN if the tensor contains any NaN values
-        // This is faster than checking each element individually
-        // because it rolls up the NaN values into a single value
-        let sum = Float::sum(self.primitive.clone());
-
-        // Check if the sum is NaN by comparing it to itself
-        Tensor::new(Float::not_equal(sum.clone(), sum))
-    }
-
-    /// Returns a new tensor with boolean elements indicating whether each element of the input is infinite (either +INF or -INF).
-    ///
-    /// # Returns
-    ///
-    /// A boolean tensor where `true` indicates that the value is infinite
-    ///
-    /// # Example
-    ///
-    /// ```rust
-    /// use burn_tensor::backend::Backend;
-    /// use burn_tensor::{Tensor, Bool, Shape};
-    ///
-    /// fn example<B: Backend>() {
-    ///    let device = B::Device::default();
-    ///    let tensor = Tensor::<B, 2>::from_data([[1.0, f64::INFINITY, 3.0], [f64::NAN, 9.0, 6.0]], &device);
-    ///    let tensor = tensor.is_finite();
-    ///    println!("{tensor}");
-    ///    // [[false, true, false], [false, false, false]]
-    /// }
-    /// ```
-    pub fn is_inf(&self) -> Tensor<B, D, Bool> {
-        self.clone()
-            .abs()
-            .equal(Tensor::full_like(self, f64::INFINITY))
-    }
-
-    /// Returns a new tensor with boolean elements indicating whether each element of the input is finite
-    ///
-    /// # Returns
-    ///
-    /// A boolean tensor where `true` indicates that the value is finite and `false` indicates
-    /// either INF, -INF or NAN
-    ///
-    /// # Example
-    ///
-    /// ```rust
-    /// use burn_tensor::backend::Backend;
-    /// use burn_tensor::{Tensor, Bool, Shape};
-    ///
-    /// fn example<B: Backend>() {
-    ///    let device = B::Device::default();
-    ///    let tensor = Tensor::<B, 2>::from_data([[1.0, f64::INFINITY, 3.0], [f64::NAN, 9.0, 6.0]], &device);
-    ///    let tensor = tensor.is_finite();
-    ///    println!("{tensor}");
-    ///    // [[true, false, true], [false, true, true]]
-    /// }
-    /// ```
-    pub fn is_finite(&self) -> Tensor<B, D, Bool> {
-        self.is_nan().bool_not().bool_and(self.is_inf().bool_not())
     }
 }
